backend/pyproject.toml
[project]
name = "talent-inbound-os"
version = "0.1.0"
description = "AI-assisted inbound recruiting management for Senior Engineers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Cristopher RL"}]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "dependency-injector>=4.45",
    "structlog>=24.4",
    "bcrypt>=4.2",
    "python-jose[cryptography]>=3.3",
    "arq>=0.26",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "sse-starlette>=2.1",
    "pypdf>=5.0",
    "python-docx>=1.1",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "factory-boy>=3.3",
    "testcontainers[postgres]>=4.9",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "B008",   # function-call-in-default-argument (FastAPI Depends() pattern)
]

[tool.ruff.lint.per-file-ignores]
"src/talent_inbound/worker.py" = ["RUF012"]  # Arq WorkerSettings uses mutable class attrs

[tool.ruff.lint.isort]
known-first-party = ["talent_inbound"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src/talent_inbound"]
omit = ["*/tests/*", "*/alembic/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
